Move circle countdown to generic ui components folder
